Publish allocated project numbers
When pressing publish button it allocates project numbers then loads all the projects again. Also more bug fixes with authentication tokens and added a redirect to that users that arent logged in are sent back to the homepage if they access a page thats not allowed. (Temporary for now as its not assigned per role). When creating a new semester it also now sets the semester id to the current semester.